requirements.txt
# Core dependencies
openai>=1.0.0
httpx>=0.24.0
aiohttp>=3.8.0
requests>=2.28.0

# Data processing
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Web framework
fastapi>=0.100.0
uvicorn>=0.22.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging
loguru>=0.7.0

# Utilities
click>=8.0.0
tqdm>=4.64.0
python-dateutil>=2.8.0

# Optional for development
pytest>=7.0.0
pytest-asyncio>=0.21.0